requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Async Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client for APIs
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Export
openpyxl==3.1.2

# WebSockets
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
ruff==0.1.6

# Environment
python-dotenv==1.0.0

# Caching
aioredis==2.0.1

# Financial data
yfinance==0.2.28

# Rate limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6